"BT ss1 "

+++ Module_2.1/src/ss1_introduction/BT/ConvertMoney.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class ConvertMoney {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int money;
+        double money;
         System.out.println("Nhập vào số tiền USD : ");
-        money = scanner.nextInt();
+        money = scanner.nextDouble();
         System.out.println(money + "USD" + " = " + 23000*money + " VND");
     }
 }
